+++ src/main/java/com/transformers/battle/transformers/services/BattleService.java
@@ -5,5 +5,5 @@ import com.transformers.battle.transformers.model.Outcome;
 import java.util.List;
 
 public interface BattleService {
-    Outcome battle(List<Integer> transformers);
+    Outcome battle(List<Long> transformers);
 }
